backend/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
sqlalchemy==2.0.38
psycopg[binary]==3.2.4
redis==5.2.1
celery==5.4.0
alembic==1.14.1
pydantic==2.10.6
pydantic-settings==2.7.1
python-multipart==0.0.20
easyocr==1.7.2
openai==1.65.1
razorpay==1.4.1
pytest==8.3.4
sentry-sdk==2.22.0
PyJWT==2.10.1
